UPDATE SilverStripe 4 compatibility
