+++ .gitattributes
@@ -1,5 +1,6 @@
-/tests export-ignore
-/docs export-ignore
-/.travis.yml export-ignore
+/tests            export-ignore
+/client/src       export-ignore
+/docs             export-ignore
+/.gitattributes   export-ignore
+/.travis.yml      export-ignore
 /.scrutinizer.yml export-ignore
-/phpunit.xml export-ignore
