+++ .scrutinizer.yml
@@ -1,31 +1,16 @@
+
 inherit: true
 
-#Copied from https://www.adayinthelifeof.nl/2013/11/20/external-code-coverage-with-travis-scrutinizer/
-tools:
-    external_code_coverage:
-      timeout: 600
-    php_code_sniffer:
-      config:
-        standard: PSR2
-    php_cs_fixer:
-      extensions:
-        # Default:
-        - php
-      fixers: []
-      enabled: false
-      filter:
-            paths: [tests/*,code/*]
-            excluded_paths: []
-coding_style:
-  php:
-    indentation:
-      general:
-        use_tabs: false
+build:
+  nodes:
+    analysis:
+      tests:
+        override: [php-scrutinizer-run]
 
 checks:
-    php:
-        code_rating: true
-        duplication: true
+  php:
+    code_rating: true
+    duplication: true
 
 filter:
-    paths: [tests/*,code/*]
+  paths: [src/*, tests/*]
