+++ LICENSE.md
@@ -1,4 +1,4 @@
-Copyright (c) 2016, Dynamic, Inc.
+Copyright (c) 2019, Dynamic, Inc. 
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
