+++ README.md
@@ -10,12 +10,33 @@ SilverStripe UploadField pre-configured for Image uploads. Allows max file size 
 
 ## Requirements
 
-- SilverStripe 3.2
+* [silverstripe/silverstripe-framework](https://github.com/silverstripe/silverstripe-framework) ^4.0
 
 ## Installation
 
-	composer require dynamic/silverstripe-imageuploadfield
+`composer require dynamic/imageuploadfield ^4.0`
 
 ## Documentation
 
 See the [docs/en](docs/en/index.md) folder.
+
+## Bugtracker
+Bugs are tracked in the issues section of this repository. Before submitting an issue please read over 
+existing issues to ensure yours is unique. 
+ 
+If the issue does look like a new bug:
+ 
+ - Create a new issue
+ - Describe the steps required to reproduce your issue, and the expected outcome. Unit tests, screenshots 
+ and screencasts can help here.
+ - Describe your environment as detailed as possible: SilverStripe version, Browser, PHP version, 
+ Operating System, any installed SilverStripe modules.
+ 
+Please report security issues to the module maintainers directly. Please don't file security issues in the bugtracker.
+
+## Maintainer Contact
+
+ *  [Dynamic](http://www.dynamicagency.com) (<dev@dynamicagency.com>)
+
+## License
+See [License](license.md)
