--- _config.php
@@ -1,4 +1,0 @@
-<?php
-
-define('IMAGEUPLOADFIELD_PATH', __DIR__);
-define('IMAGEUPLOADFIELD_DIR', basename(__DIR__));
