+++ composer.json
@@ -1,31 +1,46 @@
 {
-    "name": "dynamic/imageuploadfield",
-    "description": "SilverStripe UploadField pre-configured for Image uploads. Allows max file size to be set site wide.",
-    "authors": [
-        {
-            "name": "Dynamic, Inc.",
-            "email": "dev@dy.ag"
-        }
-    ],
-    "keywords": [
-        "silverstripe", "dynamic", "image", "upload", "field", "module"
-    ],
-    "type": "silverstripe-module",
-    "license": "BSD-3-Clause",
-    "require": {
-        "silverstripe/framework": "^4.0"
-    },
-    "require-dev": {
-        "phpunit/PHPUnit": "^3.7"
-    },
-    "config": {
-        "process-timeout": 600
-    },
-    "minimum-stability": "dev",
-    "prefer-stable": true,
-    "extra": {
+	"name": "dynamic/imageuploadfield",
+	"description": "SilverStripe UploadField pre-configured for Image uploads. Allows max file size to be set site wide.",
+	"authors": [
+		{
+			"name": "Dynamic",
+			"homepage": "http://www.dynamicagency.com",
+			"email": "dev@dynamicagency.com"
+		}
+	],
+	"keywords": [
+		"silverstripe",
+		"image",
+		"upload",
+		"field"
+	],
+	"type": "silverstripe-vendormodule",
+	"license": "BSD-3-Clause",
+	"require": {
+		"silverstripe/framework": "^4.0"
+	},
+	"require-dev": {
+		"phpunit/phpunit": "^5.7",
+		"squizlabs/php_codesniffer": "^3.0"
+	},
+	"autoload": {
+		"psr-4": {
+			"Dynamic\\ImageUpload\\": "src/"
+		}
+	},
+	"autoload-dev": {
+		"psr-4": {
+			"Dynamic\\ImageUpload\\Test\\": "tests/"
+		}
+	},
+	"config": {
+		"process-timeout": 600
+	},
+	"minimum-stability": "dev",
+	"prefer-stable": true,
+	"extra": {
 		"branch-alias": {
-      		"dev-master": "2.x-dev"
-    	}
-    }
+			"dev-master": "2.x-dev"
+		}
+	}
 }
