+++ phpcs.xml.dist
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<ruleset name="SilverStripe">
+    <description>CodeSniffer ruleset for SilverStripe coding conventions.</description>
+
+    <!-- base rules are PSR-12 -->
+    <rule ref="PSR12" >
+        <!-- Current exclusions -->
+        <!--<exclude name="PSR1.Files.SideEffects.FoundWithSymbols" />-->
+        <exclude name="PSR1.Methods.CamelCapsMethodName.NotCamelCaps" />
+    </rule>
+</ruleset>
