+++ phpunit.xml
@@ -1,12 +1,13 @@
-<phpunit backupGlobals="false" backupStaticAttributes="false" colors="true" convertErrorsToExceptions="true" convertNoticesToExceptions="true" convertWarningsToExceptions="false" processIsolation="false" stopOnFailure="false" syntaxCheck="false">
-    <testsuites>
-        <testsuite>
-            <directory suffix="Test.php">tests</directory>
-        </testsuite>
-    </testsuites>
+<phpunit bootstrap="vendor/silverstripe/cms/tests/bootstrap.php" colors="true">
+    <testsuite name="dynamic/imageuploadfield">
+        <directory>tests/</directory>
+    </testsuite>
     <filter>
         <whitelist addUncoveredFilesFromWhitelist="true">
-            <directory suffix=".php">code</directory>
+            <directory suffix=".php">src/</directory>
+            <exclude>
+                <directory suffix=".php">tests/</directory>
+            </exclude>
         </whitelist>
     </filter>
 </phpunit>
